// frontend/src/types/Employee.ts
export interface EmployeeWithId {
	_id: string;
	avatar: string;
	name: string;
	email: string;
	phone: string;
	role: string;
	position: string;
	status: string;
}

export interface Employee {
	avatar: string;
	name: string;
	email: string;
	phone: string;
	role: string;
	position: string;
	status: string;
}